[RFC] Bookkeeping: Allow merging associated income/asset or expense/liability balances
(Currently this is only a rough sketch. Especially the naming of
functions and variables is not well-thought yet. This commit is
only meant to discuss the idea.)

The idea is to calculate the "difference" between e.g. an income account
and its associated asset account. This is useful to calculate realized
income for each income (and expense) account separately as required for
EÜR in non-profits.

Here is an example:

    $ python manage.py shell_plus
    >>> mr = Account.objects.filter(name='Member fees receivable').first()
    >>> m = Account.objects.filter(name='Member fees').first()
    >>> m.associate(mr)
    >>> m.balances()
    {'debit': Decimal('0.00'), 'credit': Decimal('33822.00'), 'net': Decimal('33822.00')}
    >>> mr.balances()
    {'debit': Decimal('33822.00'), 'credit': Decimal('45.00'), 'net': Decimal('33777.00')}
    >>> m.balances_with_associated()
    {'net': Decimal('45.00'), 'debit': Decimal('0.00'), 'credit': Decimal('45.00')}